Extract getGitItems
